update querySelectors to simplify and keep pre formatting + fix search results and entityName selectors + fix error Search results popup removed +
