+++ src/activeTab/scraper/extractor/extractPages.js
@@ -1,9 +1,8 @@
-import {sleep} from "../../../common/utils";
 import {capitalizeFirst, formatLineBreaks} from "../../formatter/formatText";
 import {formatLink, setFileHeader} from "../../formatter/formatMarkdown";
-import {getPerplexityPageTitle, getPhindPageTitle, getMaxAIGooglePageTitle} from "./extractMetadata";
+import {getMaxAIGooglePageTitle, getPerplexityPageTitle, getPhindPageTitle} from "./extractMetadata";
 import {foldQuestions, unfoldQuestions} from "../../utils/webpage/interact";
-import {extractPerplexitySources} from "./extractElements";
+import {extractPerplexitySources, extractPhindSourcesFromPopup} from "./extractElements";
 
 /**
  * Exported functions
@@ -97,87 +96,60 @@ export async function extractPhindSearchPage(format) {
 }
 
 export async function extractPhindAgentPage(format) {
-  const messages = document.querySelectorAll('[name^="answer-"] > div > div');
+  const messages = document.querySelectorAll('[name^="answer-"]');
   let markdown = await setFileHeader(getPhindPageTitle(), "Phind Agent");
 
   for (const content of messages) {
-    const msgContent = content.querySelectorAll('[tabindex], div:first-of-type > div > p, div:first-of-type > div > pre, div:first-of-type > div > table');
-    const p2 = content.querySelectorAll('.card-body > div:nth-last-of-type(1) a');
-    const p3 = content.querySelectorAll('.card-body > span');
+    const msgContent = content.querySelectorAll('.col > div > div > div:has(*), .col > div > div > div:has(*):first-of-type'); //TODO
+    const searchResults = content.querySelectorAll('.col > div > div > div:nth-last-of-type(1) a');
+    const entityName = content.querySelectorAll('.col > div > div > span');
 
-    const messageText =
-      msgContent.length > 0 ? await (async () => {
-          let res = "";
+    if (msgContent.length === 0) continue;
 
-          // Extract writer name
-          if (p3.length > 0) {
-            res += "#### ";
-            let putSeparator = true;
-            p3.forEach((elt) => {
-              res += format(elt.innerHTML);
-              if (p3.length > 1 && p3[1].innerHTML !== "" && putSeparator) {
-                res += " - ";
-                putSeparator = false;
-              }
-            });
-            res += "\n";
-          }
+    let res = "";
 
-          // Extract message
-          if (p2.length > 0) // If there are search results
-          {
-            // Message
-            res += format(msgContent[0].innerHTML) + "\n";
+    // Extract writer name
+    if (entityName.length > 0) {
+      res += "#### ";
+      let putSeparator = true;
+      entityName.forEach((elt) => {
+        res += format(elt.innerHTML);
+        if (entityName.length > 1 && entityName[1].innerHTML !== "" && putSeparator) {
+          res += " - ";
+          putSeparator = false;
+        }
+      });
+      res += "\n";
+    }
 
-            // Export search results
-            res += "___\n**Sources:**";
-            const buttonsInCard = msgContent[2].querySelectorAll("button");
-            for (const btn of buttonsInCard) {
-              if (btn.textContent.toLowerCase() === "view all search results") {
-                // Open modal
-                btn.click();
-                await sleep(0); // Needed to wait for the modal to open (even if it's 0!)
+    // Extract message
+    if (searchResults.length > 0) // If there are search results
+    {
+      // Message
+      res += format(msgContent[0].innerHTML) + "\n";
 
-                // Export sources and all search results, put correct index in front of each link
-                let i = 1;
-                let allResults = "**All search results:**";
+      // Export search results
+      res += "___\n**Sources:**";
 
-                const dialogLinks = Array.from(document.querySelectorAll("[role='dialog'] a"));
-                const p2Array = Array.from(p2);
-                dialogLinks.forEach((link) => {
-                  // If the link is in the sources, add it to the sources with the correct index
-                  if (p2Array.find((elt) => elt.getAttribute("href") === link.getAttribute("href"))) {
-                    res += "\n- " + format(link.outerHTML).replace("[", `[(${i}) `);
-                  }
+      if (msgContent[2]) {
+        res = await extractPhindSourcesFromPopup(msgContent, searchResults, res, format);
+      }
+      else {
+        let i = 1;
+        searchResults.forEach((link) => {
+          res += "\n- " + format(link.outerHTML).replace("[", `[(${i}) `);
+          i ++;
+        })
+      }
 
-                  // Add the link to the all search results with the correct index
-                  allResults += "\n- " + format(link.outerHTML).replace("[", `[(${i}) `);
-                  i++;
-                });
+      res += "\n";
 
-                // Append all search results after the sources
-                res += "\n\n" + allResults;
+    } else // If there are no search results
+      msgContent.forEach((elt) => {
+        res += format(elt.innerHTML) + "\n";
+      });
 
-                // Close modal
-                document.querySelectorAll("[role='dialog'] [type='button']").forEach((btn) => {
-                  if (btn.textContent.toLowerCase() === "close") btn.click();
-                });
-              }
-            }
-
-            res += "\n";
-
-          } else // If there are no search results
-            msgContent.forEach((elt) => {
-              res += format(elt.innerHTML) + "\n";
-            });
-
-          return res;
-        })() :
-
-        '';
-
-    if (messageText !== "") markdown += messageText + "\n\n";
+    if (res !== "") markdown += res + "\n\n";
   }
 
   return markdown;
